refactor: split creative commission generation spec into a zod-free leaf (#6816)
server/lib/creativeCommissionValidation.js pulls zod, so commissionForm.js
could only mirror its per-key type/bounds/default spec by hand. That hand
copy had already drifted three ways: durationMode's default (client 'auto'
vs server 'manual'), the video backend enum (client missing fal/reactor),
and the image backend enum (client missing agy).

Move GENERATION_KEY_DEFS, ABILITY_GENERATION_SPEC, and the quality/aspect
ratio/backend enums into a new dependency-free leaf,
server/lib/creativeCommissionSpec.js (the creativeBriefLimits.js precedent).
creativeCommissionValidation.js now imports and re-exports them, still
building the same Zod schema from GENERATION_KEY_DEFS.

Also resolves the durationMode compatibility split explicitly: the
descriptor now carries both `default` ('auto', the fresh-commission seed
from #4494) and `legacyAbsent` ('manual', what a record with no key at all
has always meant). abilityAdapters.js's coerceGenerationValue and its three
raw duration reads now resolve an absent value through this split
(resolvedDefault / GENERATION_KEY_DEFS.durationMode.legacyAbsent) instead of
a bare `.default`, so stored-record behavior is unchanged.
